TrajectoryOptimization.jl

TrajectoryOptimization.jl

Documentation for TrajectoryOptimization.jl

TrajectoryOptimization.jl

Overview

Getting Started

Creating a Model

Overview

The purpose of this package is to provide a testbed for state-of-the-art trajectory optimization algorithms. In general, this package focuses on trajectory optimization problems of the form (put LaTeX here)

This package currently implements both indirect and direct methods for trajectory optimization:

Iterative LQR (iLQR): indirect method based on differential dynamic programming

Direct Collocation: direct method that formulates the problem as an NLP and passes the problem off to a commercial NLP solver

The primary focus of this package is developing the iLQR algorithm, although we hope this will extend to many algorithms in the future.

Getting Started

In order to set up a trajectory optimization problem, the user needs to create a Model and Objective

Creating a Model

There are two ways of creating a model:

An in-place analytic function of the form f(ẋ,x,u)

A URDF

Analytic Models

To create an analytic model we just need to

TrajectoryOptimization.Solver — Type.

struct Solver{M<:Model, O<:TrajectoryOptimization.Objective}

Type for solver

source

TrajectoryOptimization.Model — Type.

Model(mech::Mechanism, torques::Array{Bool, 1})

Constructor for an underactuated mechanism, where torques is a binary array that specifies whether a joint is actuated.

source

Model(urdf)

Construct a fully actuated model from a string to a urdf file

source

Model(urdf, torques)

Construct a partially actuated model from a string to a urdf file

source

TrajectoryOptimization.SolverResults — Type.

abstract type SolverResults

Abstract type for the output of solving a trajectory optimization problem

source