Setting up a Dynamics Model

Setting up a Dynamics Model

Documentation for TrajectoryOptimization.jl

TrajectoryOptimization.jl

Overview

Getting Started

Creating a Model

Creating an Objective

Solving the Problem

Solve Methods

DIRCOL Method

Constraints

Constraint Type

Creating Constraints

Setting up a Dynamics Model

Overview

Analytical Models

Overview

The Model type holds information about the dynamics of the system. All dynamics are assumed to be state-space models of the system of the form ẋ = f(x,u) where ẋ is the state derivative, x an n-dimentional state vector, and u in an m-dimensional control input vector. The function f can be any nonlinear function.

TrajectoryOptimization.jl poses the trajectory optimization problem by discretizing the state and control trajectories, which requires discretizing the dynamics, turning the continuous time differential equation into a discrete time difference equation of the form x[k+1] = f(x[k],u[k]), where k is the time step. There many methods of performing this discretization, and TrajectoryOptimization.jl offers several of the most common methods.

Sometimes is it convenient to write down the difference equation directly, rather than running a differential equation through a discretizing integration method. TrajectoryOptimization.jl offers method deal directly with either continuous differential equations, or discrete difference equations.

The Model type is parameterized by the DynamicsType, which is either Continuous, or Discrete. The models holds the equation f and it's Jacobian, ∇f, along with the dimensions of the state and control vectors.

Models can be created by writing down the dynamics analytically or be generated from a URDF file via RigidBodyDynamics.jl.

Analytical Models

TrajectoryOptimization.AnalyticalModel — Type.

struct AnalyticalModel{D} <: Model{D}

Dynamics model

Holds all information required to uniquely describe a dynamic system, including a general nonlinear dynamics function of the form ẋ = f(x,u), where x ∈ ℜⁿ are the states and u ∈ ℜᵐ are the controls.

Dynamics function Model.f should be in the following forms: 'f!(ẋ,x,u)' and modify ẋ in place

source

Continuous Models

Continuous models assume differential equations are specified by an in-place function in one of the following forms:

f!(ẋ,x,u)
f!(ẋ,x,u,p)

and Jacobians of the form

∇f!(Z,x,u)
∇f!(Z,x,u,p)

where ẋ is the state derivative, p is a NamedTuple of model parameters, and Zis the (n × (n+m)) Jacobian matrix (i.e. [∇ₓf(x,u) ∇ᵤf(x,u)]).

The following constructors can be used to create Continuous Analytical models

Model(::Function,::Int,::Int,::Dict)
Model(::Function,::Int,::Int,::NamedTuple,::Dict)
Model(::Function,::Function,::Int,::Int,::Dict)
Model(::Function,::Function,::Int,::Int,::NamedTuple,::Dict)