Converting to an NLP

Converting to an NLP

Important Types

The TrajOptNLP type

MathOptInterface

Trajectory optimization problems are really just nonlinear programs (NLPs). A handful of high-quality NLP solvers exist, such as Ipopt, Snopt, and KNITRO. TrajectoryOptimization provides an interface that allows methods that are amenable to use with a general-purpose NLP solver. In the NLP, the states and constraints at every knot point are concatenated into a single large vector of decision variables, and the cost hessian and constraint Jacobians are represented as large, sparse matrices.

Important Types

Below is the documentation for the types used to represent a trajectory optimization problem as an NLP:

Missing docstring.

Missing docstring for NLPData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NLPConstraintSet. Check Documenter's build log for details.

Missing docstring.

Missing docstring for QuadraticViewCost. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ViewKnotPoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TrajData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NLPTraj. Check Documenter's build log for details.

The TrajOptNLP type

The most important type is the TrajOptNLP, which is a single struct that has all the required methods to evaluate the trajectory optimization problem as an NLP.

Missing docstring.

Missing docstring for TrajOptNLP. Check Documenter's build log for details.

Interface

Use the following methods on a TrajOptNLP nlp. Unless otherwise noted, Z is a single vector of NN decision variables (where NN is the total number of states and controls across all knot points).

Missing docstring.

Missing docstring for eval_f. Check Documenter's build log for details.

Missing docstring.

Missing docstring for grad_f!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hess_f!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hess_f_structure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eval_c!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jac_c!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jacobian_structure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hess_L!. Check Documenter's build log for details.

The following methods are useful to getting important information that is typically required by an NLP solver

TrajectoryOptimization.primal_bounds! — Function

primal_bounds!(zL, zU, con::AbstractConstraint)

Set the lower zL and upper zU bounds on the primal variables imposed by the constraint con. Return whether or not the vectors zL or zU could be modified by con (i.e. if the constraint con is a bound constraint).

primal_bounds!(zL, zU, cons::ConstraintList; remove=true)

Get the lower and upper bounds on the primal variables imposed by the constraints in cons, where zL and zU are vectors of length NN, where NN is the total number of primal variables in the problem. Returns the modified lower bound zL and upper bound zU.

If any of the bound constraints are redundant, the strictest bound is returned.

If remove = true, these constraints will be removed from cons.

Missing docstring.

Missing docstring for constraint_type. Check Documenter's build log for details.

MathOptInterface

The TrajOptNLP can be used to set up an MathOptInterface.AbstractOptimizer to solve the trajectory optimization problem. For example if we want to use Ipopt and have already set up our TrajOptNLP, we can solve it using build_MOI!(nlp, optimizer):

using Ipopt
using MathOptInterface
nlp = TrajOptNLP(...)  # assume this is already set up
optimizer = Ipopt.Optimizer()
TrajectoryOptimization.build_MOI!(nlp, optimizer)
MathOptInterface.optimize!(optimizer)