3. Creating Constraints

3. Creating Constraints

Creating Constraint Sets

Defined Constraints

Constraint Type

Adding a New Constraint

API

Creating Constraint Sets

The easiest way to set up the constraints for your problem is through the ConstraintSet. This structure simply holds a vector of all the constraints in the trajectory optimization problem. The easiest way to start is to create an empty ConstraintSet:

conSet = ConstraintSet(n,m,N)

Adding Constraints

You can add any constraint (the list of currently implemented constraints is given in the following section) to the constraint set using the following method:

TrajectoryOptimization.add_constraint! — Function

add_constraint!(conSet, conVal::ConstraintVals, idx=-1)
add_constraint!(conSet, con::AbstractConstraint, inds::UnitRange, idx=-1)

Add a constraint to the constraint set. You can directly add a ConstraintVals type, but when adding a normal AbstractConstraint you must specify the range of knotpoints to which the constraint applies.

The optional idx argument allows you to specify exactly where in the vector of constraints you'd like the constraint to be added. The ordering will effect the order in which the constraint appears in concatenated vectors of constraint values (that show up in direct methods), potentially effecting the band structure of the resultant Jacobian. See solvers for more details.

Other Methods

There are a variety of methods defined on constraint sets, since they are the general interface that solvers work with when dealing with constraints.

Information Methods

These methods provide or calculate information about the constraint set

Base.size — Method

Get size of state and control dimensions

Base.length — Method

Get number of separate constraints (i.e. ConstraintVals) in the set

TrajectoryOptimization.num_constraints! — Function

Re-calculate the number of constraints in the constraint set

TrajectoryOptimization.num_constraints — Function

num_constraints(::ConstraintSet)
num_constraints(::AbstractSolver)
num_constraints(::Problem)

Get the total number of constraints at each time step

Calculation Methods

These methods perform calculations on the constraint set

TrajectoryOptimization.max_violation — Function

max_violation(conSet::ConstraintSet)
max_violation(conSet::ConstraintSet, Z::Traj)
max_violation(prob::Problem, Z=prob.Z)
max_violation(solver::AbstractSolver)
max_violation(solver::AbstractSolver, Z)

Calculate the maximum constraint violation for the entire constraint set. If the a trajectory is not passed in, the violation is computed from the currently stored constraint values; otherwise, the constraints are re-computed using the trajectory passed in.

TrajectoryOptimization.max_penalty — Function

Calculate the maximum penalty parameter across all constraints

TrajectoryOptimization.evaluate! — Method

evaluate!(conSet::ConstraintSet, Z::Traj)

Compute constraint values for all constraints for the entire trajectory

TrajectoryOptimization.jacobian! — Method

jacobian!(conSet::ConstraintSet, Z::Traj)

Compute constraint Jacobians for all constraints for the entire trajectory

TrajectoryOptimization.reset! — Method

Reset all the Lagrange multipliers and constraint values to zero and penalties the their initial value

ConstraintSet supports indexing and iteration, which returns the ConstraintVals at that index. However, to avoid allocations, iteration directly on the .constraints field.

Additionally, to avoid allocations when computing max_violation, you can call max_violation!(conSet) and then maximum(conSet.c_max) to perform the reduction in the scope where the result is stored (thereby avoiding an allocation).

Defined Constraints

The following constraints are currently defined. See later sections on the constraint interface and how to add custom constraints.

TrajectoryOptimization.GoalConstraint — Type

struct GoalConstraint{T, P, N, L} <: TrajectoryOptimization.AbstractConstraint{Equality,State,P}

Constraint of the form $x_g = a$, where $x_g$ can be only part of the state vector.

Constructors:

GoalConstraint(xf::AbstractVector)
GoalConstraint(xf::AbstractVector, inds)

where xf is an n-dimensional goal state. If inds is provided, only xf[inds] will be used.

TrajectoryOptimization.BoundConstraint — Type

struct BoundConstraint{T, P, PN, NM, PNM} <: TrajectoryOptimization.AbstractConstraint{Inequality,Stage,P}

Linear bound constraint on states and controls

Constructors

BoundConstraint(n, m; x_min, x_max, u_min, u_max)

Any of the bounds can be ±∞. The bound can also be specifed as a single scalar, which applies the bound to all state/controls.

TrajectoryOptimization.CircleConstraint — Type

struct CircleConstraint{T, P} <: TrajectoryOptimization.AbstractConstraint{Inequality,State,P}

Constraint of the form $(x - x_c)^2 + (y - y_c)^2 \leq r^2$ where $x$, $y$ are given by x[xi],x[yi], $(x_c,y_c)$ is the center of the circle, and $r$ is the radius.

Constructor: ```julia

CircleConstraint(n, xc::SVector{P}, yc::SVector{P}, radius::SVector{P}, xi=1, yi=2) ```

TrajectoryOptimization.SphereConstraint — Type

struct SphereConstraint{T, P} <: TrajectoryOptimization.AbstractConstraint{Inequality,State,P}

Constraint of the form $(x - x_c)^2 + (y - y_c)^2 + (z - z_c)^2 \leq r^2$ where $x$, $y$, $z$ are given by x[xi],x[yi],x[zi], $(x_c,y_c,z_c)$ is the center of the sphere, and $r$ is the radius.

Constructor:

SphereConstraint(n, xc::SVector{P}, yc::SVector{P}, zc::SVector{P},
	radius::SVector{P}, xi=1, yi=2, zi=3)

TrajectoryOptimization.NormConstraint — Type

struct NormConstraint{S, W<:Union{Control, State}, T} <: TrajectoryOptimization.AbstractConstraint{S,W<:Union{Control, State},1}

Constraint of the form $\norm(y)^2 \leq val$ where $y$ is either a state or a vector (but not both)

Constructors:

NormConstraint{S,State}(n,val)
NormConstraint{S,Control}(m,val)

TrajectoryOptimization.IndexedConstraint — Type

struct IndexedConstraint{S, W, P, N, M, NM, Bx, Bu, C} <: TrajectoryOptimization.AbstractConstraint{S,W,P}

Compute a constraint on an arbitrary portion of either the state or control, or both. Useful for dynamics augmentation. e.g. you are controlling two models, and have individual constraints on each. You can define constraints as if they applied to the individual model, and then wrap it in an IndexedConstraint to apply it to the appropriate portion of the concatenated state. Assumes the indexed state portion is contiguous.

Type params:

S - Inequality or Equality

W - ConstraintType

P - Constraint length

N,M - original state and control dimensions

NM - N+M

Bx - location of the first element in the state index

Bu - location of the first element in the control index

C - type of original constraint

Constructors:julia IndexedConstraint(n, m, con) IndexedConstraint(n, m, con, ix::SVector, iu::SVector) where the arguments n and m are the state and control dimensions of the new dynamics. ix and iu are the indices into the state and control vectors. If left out, they are assumed to start at the beginning of the vector.

NOTE: Only part of this functionality has been tested. Use with caution!

Constraint Type

All constraints inherit from AbstractConstraint{S<:ConstraintSense,W<:ConstraintType,P}, where ConstraintSense specifies Inequality or Equality, ConstraintType specifies the "bandedness" of the constraint (will be discussed more later), and P is the dimension of the constraint. This allows the software to easily dispatch over the type of constraint. Each constraint type represents a vector-valued constraint. The intention is that each constraint type represent one line in the constraints of problem definition (where they may be vector or scalar-valued).

TrajectoryOptimization.jl assumes equality constraints are of the form $g(x) = 0$ and inequality constraints are of the form $h(x) \leq 0$.

TrajectoryOptimization.AbstractConstraint — Type

abstract type AbstractConstraint <: TrajectoryOptimization.GeneralConstraint

Abstract vector-valued constraint of size P for a trajectory optimization problem. May be either inequality or equality (specified by S<:ConstraintSense), and be function of single, adjacent, or all knotpoints (specified by W<:ConstraintType).

Interface: Any constraint type must implement the following interface:

n = state_dim(::MyCon)
m = control_dim(::MyCon)
p = Base.length(::MyCon)
c = evaluate(::MyCon, args...)   # args determined by W
∇c = jacobian(::MyCon, args...)  # args determined by W

The evaluate and jacobian (identical signatures) methods should have the following signatures

W <: State: evaluate(::MyCon, x::SVector)

W <: Control: evaluate(::MyCon, u::SVector)

W <: Stage: evaluate(::MyCon, x, u)

W <: Dynamical: evaluate(::MyCon, x′, x, u)

W <: Coupled: evaluate(::MyCon, x′, u′ x, u)

Or alternatively,

W <: Stage: evaluate(::MyCon, z::KnotPoint)

W <: Coupled: evaluate(::MyCon, z′::KnotPoint, z::KnotPoint)

The Jacobian method for State or Control is optional, since it will be automatically computed using ForwardDiff. Automatic differentiation for other types of constraints is not yet supported.

For W <: State, control_dim(::MyCon) doesn't need to be defined. Equivalently, for W <: Control, state_dim(::MyCon) doesn't need to be defined.

For W <: General, the more general evaluate and jacobian methods must be used

evaluate!(vals::Vector{<:AbstractVector}, ::MyCon, Z::Traj, inds=1:length(Z)-1)
jacobian!(∇c::Vector{<:AbstractMatrix}, ::MyCon, Z::Traj, inds=1:length(Z)-1)

These methods can be specified for any constraint, instead of the not-in-place functions above.

TrajectoryOptimization.ConstraintSense — Type

Specifies whether the constraint is an equality or inequality constraint. Valid subtypes are Equality, Inequality, and Null

TrajectoryOptimization.ConstraintType — Type

Specifies the ``bandedness'' of the constraint. This ends up being the width of the constraint Jacobian, or the total number of input variables. This is important to reduce the size of arrays needed to store the Jacobian, as well as special-case the block matrix algebra.

Current subtypes:

Stage

State <: Stage

Control <: Stage

Coupled

Dynamical <: Coupled

CoupledState <: Coupled

CoupledControl <: Coupled

General

GeneralState <: General

GeneralControl <: General

Methods

The following methods are defined for all AbstractConstraints

TrajectoryOptimization.state_dims — Function

Size of control vector

TrajectoryOptimization.control_dims — Function

Size of state vector

TrajectoryOptimization.evaluate! — Function

evaluate!(vals::Vector{<:AbstractVector}, con::AbstractConstraint{S,W,P},
	Z, inds=1:length(Z)-1)

Evaluate constraints for entire trajectory. This is the most general method used to evaluate constraints, and should be the one used in other functions.

For W<:Stage this will loop over calls to evaluate(con,Z[k])

For W<:Coupled this will loop over calls to evaluate(con,Z[k+1],Z[k])

For W<:General,this must function must be explicitly defined. Other types may define it if desired.

evaluate!(conSet::ConstraintSet, Z::Traj)

Compute constraint values for all constraints for the entire trajectory

TrajectoryOptimization.jacobian! — Function

jacobian!(vals::Vector{<:AbstractVector}, con::AbstractConstraint{S,W,P},
	Z, inds=1:length(Z)-1)

Evaluate constraint Jacobians for entire trajectory. This is the most general method used to evaluate constraint Jacobians, and should be the one used in other functions.

For W<:Stage this will loop over calls to jacobian(con,Z[k])

For W<:Coupled this will loop over calls to jacobian(con,Z[k+1],Z[k])

For W<:General,this must function must be explicitly defined. Other types may define it if desired.

jacobian!(conSet::ConstraintSet, Z::Traj)

Compute constraint Jacobians for all constraints for the entire trajectory

TrajectoryOptimization.contype — Function

Get type of constraint (bandedness)

TrajectoryOptimization.sense — Function

Get constraint sense (Inequality vs Equality)

TrajectoryOptimization.width — Function

Returns the width of the constraint Jacobian, i.e. the total number of inputs to the constraint

TrajectoryOptimization.upper_bound — Function

Upper bound of the constraint, as a vector, which is 0 for all constraints (except bound constraints)

TrajectoryOptimization.lower_bound — Function

Upper bound of the constraint, as a vector, which is 0 equality and -Inf for inequality (except bound constraints)

TrajectoryOptimization.is_bound — Function

Is the constraint a bound constraint or not

TrajectoryOptimization.check_dims — Function

Check whether the constraint is consistent with the specified state and control dimensions

Adding a New Constraint

See interface description in documentation for AbstractConstraint. The interface allows for a lot of flexibility, but let's do a simple example. Let's say we have a 2-norm constraint on the controls at each time step, e.g. $||u|| \leq a$. We can do this with just a few lines of code:

struct ControlNorm{T} <: AbstractConstraint{Inequality,Control,1}
  m::Int
  a::T
end
control_dim(con::ControlNorm) = con.m
evaluate(con::ControlNorm, u::SVector) = @SVector [norm(u) - con.a] # needs to be a vector output
jacobian(con::ControlNorm, u::SVector) = u'/norm(u)  # optional

Importantly, note that the inheritance specifies the constraint applies only to individual controls, the constraint in an inequality, and has dimension 1.

Let's say the bound $a$ varied by time-step. We could handle this easily by instead defining the methods operating on the entire trajectory:

struct ControlNorm2{T} <: AbstractConstraint{Inequality,Control,1}
  m::Int
  a::Vector{T}
end
control_dim(con::ControlNorm) = con.m
function evaluate!(vals::Vector{<:AbstractVector}, con::ControlNorm,
    Z, inds=1:length(Z)-1)
  for (i,k) in enumerate(inds)
    u = control(Z[k])
    vals[i] = @SVector [norm(u) - con.a[k]]
  end
end
function jacobian!(∇c::Vector{<:AbstractMatrix}, con::ControlNorm,
    Z, inds=1:length(Z)-1)
  for (i,k) in enumerate(inds)
    u = control(Z[k])
    ∇c[i] = u'/norm(u)
  end
end

API

TrajectoryOptimization.ConstraintVals — Type

struct ConstraintVals{T, W, C, P, NM, PNM}

Struct that stores all of the values associated with a particular constraint. Importantly, ConstraintVals stores the list of knotpoints to which the constraint is applied. This type should be fairly transparent to the user, and only needs to be directly dealt with when writing solvers or setting fine-tuned updates per constraint (via the .params field).

TrajectoryOptimization.Stage — Type

Only a function of states and controls at a single knotpoint

TrajectoryOptimization.State — Type

Only a function of states at a single knotpoint

TrajectoryOptimization.Control — Type

Only a function of controls at a single knotpoint

TrajectoryOptimization.Coupled — Type

Only a function of states and controls at two adjacent knotpoints

TrajectoryOptimization.Dynamical — Type

Only a function of states and two adjacent knotpoints, and the control at the previous knotpoint, i.e. f(x,u) - x′

TrajectoryOptimization.CoupledState — Type

Only a function of states at adjacent knotpoints

TrajectoryOptimization.CoupledControl — Type

Only a function of controls at adjacent knotpoints

TrajectoryOptimization.General — Type

A function of all states and controls in the trajectory

TrajectoryOptimization.GeneralState — Type

A function of all states in the trajectory

TrajectoryOptimization.GeneralControl — Type

A function of all controls in the trajectory