Cost Functions and Objectives

This page details the functions related to building and evaluating cost functions and objectives.

Quadratic Cost Functions

TrajectoryOptimization.QuadraticCost — Type

mutable struct QuadraticCost{TQ, TR, TH, Tq, Tr, T} <: CostFunction

Cost function of the form 1/2xₙᵀ Qf xₙ + qfᵀxₙ + ∫ ( 1/2xᵀQx + 1/2uᵀRu + xᵀHu + q⁠ᵀx rᵀu ) dt from 0 to tf R must be positive definite, Q and Qf must be positive semidefinite

Constructor use any of the following constructors:

QuadraticCost(Q, R, H, q, r, c)
QuadraticCost(Q, R; H, q, r, c)
QuadraticCost(Q, q, c)

Any optional or omitted values will be set to zero(s).

TrajectoryOptimization.LQRCost — Function

LQRCost(Q, R, xf)

Cost function of the form $(x-x_f)^T Q (x_x_f) + u^T R u$ R must be positive definite, Q must be positive semidefinite

TrajectoryOptimization.LQRCostTerminal — Function

LQRCostTerminal(Qf, xf)

Cost function of the form $(x-x_f)^T Q (x_x_f)$ Q must be positive semidefinite

TrajectoryOptimization.LQRObjective — Function

LQRObjective(Q, R, Qf, xf, N)

Create an objective of the form $(x_N - x_f)^T Q_f (x_N - x_f) + \sum_{k=0}^{N-1} (x_k-x_f)^T Q (x_k-x_f) + u_k^T R u_k$

CostExpansion Type

The CostExpansion type stores the pieces of the second order Taylor expansion of the cost for the entire trajectory, stored as vectors of Static Vectors or Static Matrices. e.g. to get the Hessian with respect to x at knotpoint 5 you would use E.xx[5].

TrajectoryOptimization.CostExpansion — Type

struct CostExpansion{T, N, M, L1, L2, L3}

Store the terms of the 2nd order expansion for the entire trajectory

Objective

TrajectoryOptimization.Objective — Type

struct Objective{C} <: TrajectoryOptimization.AbstractObjective

Objective: stores stage cost(s) and terminal cost functions

Constructors:

Objective(cost, N)
Objective(cost, cost_term, N)
Objective(costs::Vector{<:CostFunction}, cost_term)
Objective(costs::Vector{<:CostFunction})

TrajectoryOptimization.cost — Method

cost(obj::Objective, Z::Traj)::Float64
cost(obj::Objective, dyn_con::DynamicsConstraint{Q}, Z::Traj)

Evaluate the cost for a trajectory. Calculate the cost gradient for an entire trajectory. If a dynamics constraint is given, use the appropriate integration rule, if defined.

TrajectoryOptimization.stage_cost — Function

stage_cost(cost::CostFunction, z::KnotPoint) -> Any

Evaluate the cost at a knot point

TrajectoryOptimization.get_J — Function

Get the vector of costs at each knot point. sum(get_J(obj)) is equal to the cost

TrajectoryOptimization.cost_gradient — Function

Qx,Qu = cost_gradient(cost::CostFunction, z::KnotPoint)

Get Qx, Qu pieces of gradient of cost function, multiplied by dt

TrajectoryOptimization.cost_gradient! — Function

cost_gradient!(E::CostExpansion, obj::Objective, Z::Traj)
cost_gradient!(E::CostExpansion, obj::Objective, dyn_con::DynamicsConstraint{Q}, Z::Traj)

Calculate the cost gradient for an entire trajectory. If a dynamics constraint is given, use the appropriate integration rule, if defined.

TrajectoryOptimization.cost_hessian — Function

Qxx,Quu,Qux = cost_hessian(cost::CostFunction, z::KnotPoint)

Get Qxx, Quu, Qux pieces of Hessian of cost function, multiplied by dt

TrajectoryOptimization.cost_hessian! — Function

cost_hessian!(E::CostExpansion, obj::Objective, Z::Traj)

Calculate the cost Hessian for an entire trajectory