2. Setting up an Objective

2. Setting up an Objective

Overview

Cost functions

Objectives

API

Overview

All trajectory optimization problems require a cost function at each stage of the trajectory. Cost functions must be scalar-valued. We assume general cost functions of the form,

\[\ell_f(x_N) + \sum_{k=1}^{N-1} \ell_k(x_k,u_k) dt\]

It is very important to note that $\ell_k(x_k,u_k)$ is ONLY a function of $x_k$ and $u_k$, i.e. no coupling across time-steps is permitted. This is a requirement for Differential Dynamic Programming methods such as iLQR, but could be relaxed for methods that parameterize both states and controls, such as DIRCOL. In general, any coupling between adjacent time-steps can be resolved by augmenting the state and defining the appropriate dynamics (this is the method we use to solve minimum time problems).

In general, trajectory optimization will take a second order Taylor series approximation of the cost function, resulting in a quadratic cost function of the form

\[x_N^T Q_f x_N + q_f^T x_N + \sum_{k=1}^{N-1} x_k^T Q_k x_k + q_k^T x_k + u_k^T R_k u_k + r_k^T u_k + u_k^T H_k x_k\]

This type of quadratic cost is typical for trajectory optimization problems, especially when Q is positive semi-definite and R is positive definite, which is strictly convex. These problem behave well and reduce the computational requirements of taking second-order Taylor series expansions of the cost at each iteration.

In TrajectoryOptimization.jl we differentiate between the entire objective and the cost functions at each time step. We use Objective to describe the function that is being minimized, which typically consists of a sum of cost functions, with potentially some additional terms (as is the case with augmented Lagrangian objectives). Describing the Objective as a sum of individual functions allows the solvers to more efficiently compute the gradient and Hessian of the entire cost, which is block-diagonal given the Markovianity of the problem.

Cost functions

There are several different cost function types that all inherit from CostFunction. The following sections detail the various methods for instantiating these cost function types.

Quadratic Costs

Quadratic costs are the most standard cost function and excellent place to start. Let's assume we are creating an LQR tracking cost of the form

\[(x_N - x_f)^T Q_f (x_N - x_f) + \sum_{k=1}^{N-1} (x_k - x_f)^T Q (x_k - x_f) + u_k^T R u_k\]

for the simple pendulum with the goal of doing a swing-up. To do this we have very convenient constructors LQRCost and LQRCostTerminal:

using LinearAlgebra
n,m = 2,1
Q = Diagonal(0.1I,n)
R = Diagonal(0.1I,m)
Qf = Diagonal(1000I,n)
xf = [π,0]
costfun = LQRCost(Q,R,Qf)
costfun_term = LQRCostTerminal(Qf,xf)

It is HIGHLY recommended to specify any special structure, such as Diagonal, especially since these matrices are almost always diagonal.

This constructor actually does a simple conversion to turn our cost function into a generic quadratic cost function. We could do this ourselves:

H = zeros(m,n)
q = -Q*xf
r = zeros(m)
c = xf'Q*xf/2
qf = -Qf*xf
cf = xf'Qf*xf/2
costfun      = QuadraticCost(Q, R, H, q, r, c)
costfun_term = QuadraticCost(Qf, R*0, H, qf, r*0, cf)

The QuadraticCost constructor also supports keyword arguments and one that allows for only Q,q and c.:

costfun      = QuadraticCost(Q, R, q=q, c=c)
costfun_term = QuadraticCost(Q, q, c)

Once we have defined the cost function, we can create an objective for our problem by simply copying over all time steps (except for the terminal).

# Create an objective from a single cost function
N = 51
obj = Objective(costfun, costfun_term, N)

There's also a convenient constructor that builds an LQRObjective

obj = LQRObjective(Q, R, Qf, xf, N)

QuadraticCost API

TrajectoryOptimization.QuadraticCost — Type

mutable struct QuadraticCost{TQ, TR, TH, Tq, Tr, T} <: CostFunction

Cost function of the form 1/2xₙᵀ Qf xₙ + qfᵀxₙ + ∫ ( 1/2xᵀQx + 1/2uᵀRu + xᵀHu + q⁠ᵀx rᵀu ) dt from 0 to tf R must be positive definite, Q and Qf must be positive semidefinite

Constructor use any of the following constructors:

QuadraticCost(Q, R, H, q, r, c)
QuadraticCost(Q, R; H, q, r, c)
QuadraticCost(Q, q, c)

Any optional or omitted values will be set to zero(s).

TrajectoryOptimization.LQRCost — Function

LQRCost(Q, R, xf)

Cost function of the form $(x-x_f)^T Q (x_x_f) + u^T R u$ R must be positive definite, Q must be positive semidefinite

TrajectoryOptimization.LQRCostTerminal — Function

LQRCostTerminal(Qf, xf)

Cost function of the form $(x-x_f)^T Q (x_x_f)$ Q must be positive semidefinite

TrajectoryOptimization.LQRObjective — Function

LQRObjective(Q, R, Qf, xf, N)

Create an objective of the form $(x_N - x_f)^T Q_f (x_N - x_f) + \sum_{k=0}^{N-1} (x_k-x_f)^T Q (x_k-x_f) + u_k^T R u_k$

Cost Function Interface

All cost functions are required to define the following methods

n = state_dim(cost)
m = control_dim(cost)
J = stage_cost(cost, x, u)
J = stage_cost(cost, xN)
Qx,Qu = gradient(cost, x, u)
Qxx,Quu,Qux = hessian(cost, x, u)

and inherit from CostFunction.

CostExpansion Type

The CostExpansion type stores the pieces of the second order Taylor expansion of the cost for the entire trajectory, stored as vectors of Static Vectors or Static Matrices. e.g. to get the Hessian with respect to x at knotpoint 5 you would use E.xx[5].

Objectives

Constructors

Objectives can be created by copying a single cost function over all time steps

Objective(cost::CostFunction, N::Int)

or uniquely specifying the terminal cost function

Objective(cost::CostFunction, cost_terminal::CostFunction, N::Int)

or by explicitly specifying a list of cost functions

Objective(costfuns::Vector{<:CostFunction})

Methods

Constraints extends the methods on CostFunction to the whole trajectory

cost(obj, Z)
cost_expansion!(E::CostExpansion, obj, Z)

where Z is a Trajectory (e.g. Vector of KnotPoints)

API

TrajectoryOptimization.cost — Function

cost(obj::Objective, Z::Traj)::Float64
cost(obj::Objective, dyn_con::DynamicsConstraint{Q}, Z::Traj)

Evaluate the cost for a trajectory. Calculate the cost gradient for an entire trajectory. If a dynamics constraint is given, use the appropriate integration rule, if defined.

cost(::Problem)
cost(::AbstractSolver)

Compute the cost for the current trajectory

TrajectoryOptimization.stage_cost — Function

stage_cost(cost::CostFunction, z::KnotPoint) -> Any

Evaluate the cost at a knot point

TrajectoryOptimization.get_J — Function

Get the vector of costs at each knot point. sum(get_J(obj)) is equal to the cost

TrajectoryOptimization.cost_gradient — Function

Qx,Qu = cost_gradient(cost::CostFunction, z::KnotPoint)

Get Qx, Qu pieces of gradient of cost function, multiplied by dt

TrajectoryOptimization.cost_gradient! — Function

cost_gradient!(E::CostExpansion, obj::Objective, Z::Traj)
cost_gradient!(E::CostExpansion, obj::Objective, dyn_con::DynamicsConstraint{Q}, Z::Traj)

Calculate the cost gradient for an entire trajectory. If a dynamics constraint is given, use the appropriate integration rule, if defined.

TrajectoryOptimization.cost_hessian — Function

Qxx,Quu,Qux = cost_hessian(cost::CostFunction, z::KnotPoint)

Get Qxx, Quu, Qux pieces of Hessian of cost function, multiplied by dt

TrajectoryOptimization.cost_hessian! — Function

cost_hessian!(E::CostExpansion, obj::Objective, Z::Traj)

Calculate the cost Hessian for an entire trajectory