Setting up a Dynamics Model

Setting up a Dynamics Model

Constraints

Constraint Type

Creating Constraints

Special Constraints

Setting up a Dynamics Model

Overview

Continuous Models

From analytical function

With analytical Jacobians

URDF Files

A note on Model types

Discrete Models

From a continuous model

Analytical

From Continuous Model

API

Solvers

Iterative LQR (iLQR)

Setting up a Problem

Creating a Problem

Adding constraints

TrajectoryOptimization.jl

Overview

Getting Started

Creating a Model

Creating an Objective

Solving the Problem

Solve Methods

DIRCOL Method

Overview

The Model type holds information about the dynamics of the system. All dynamics are assumed to be state-space models of the system of the form ẋ = f(x,u) where ẋ is the state derivative, x an n-dimentional state vector, and u in an m-dimensional control input vector. The function f can be any nonlinear function.

TrajectoryOptimization.jl poses the trajectory optimization problem by discretizing the state and control trajectories, which requires discretizing the dynamics, turning the continuous time differential equation into a discrete time difference equation of the form x[k+1] = f(x[k],u[k]), where k is the time step. There many methods of performing this discretization, and TrajectoryOptimization.jl offers several of the most common methods.

Sometimes is it convenient to write down the difference equation directly, rather than running a differential equation through a discretizing integration method. TrajectoryOptimization.jl offers method deal directly with either continuous differential equations, or discrete difference equations.

The Model type is parameterized by the DynamicsType, which is either Continuous, or Discrete. The models holds the equation f and it's Jacobian, ∇f, along with the dimensions of the state and control vectors.

Models can be created by writing down the dynamics analytically or be generated from a URDF file via RigidBodyDynamics.jl.

Continuous Models

From analytical function

Let's start by writing down a dynamics function for a simple pendulum with state [θ; ω] and a torque control input

function pendulum_dynamics!(xdot,x,u)
    m = 1.
    l = 0.5
    b = 0.1
    lc = 0.5
    J = 0.25
    g = 9.81
    xdot[1] = x[2]
    xdot[2] = (u[1] - m*g*lc*sin(x[1]) - b*x[2])/J
end

Note that the function is in-place, in that it writes the result to the first argument. It is also good practice to concretely specify the location to write to rather than using something like xdot[1:end] or xdot[:].

Notice that we had to specify a handful of constants when writing down the dynamics. We could have initialized them outside the scope of the function (which may result in global variables, so be careful!) or we can pass them in as a NamedTuple of parameters:

function pendulum_dynamics_params!(xdot,x,u,p)
    xdot[1] = x[2]
    xdot[2] = (u[1] - p.m * p.g * p.lc * sin(x[1]) - p.b*x[2])/p.J
end

We can now create our model using our analytical dynamics function with or without the parameters tuple

n,m = 2,1
model = Model(pendulum_dynamics!, n, m)

params = (m=1, l=0.5, b=0.1, lc=0.5, J=0.25, g=9.81)
model = Model(pendulum_dynamics_params!, n, m, params)

With analytical Jacobians

Since we have a very simple model, writing down an analytical expression of the Jacobian is pretty easy:

function pendulum_jacobian!(Z,x,u)
    m = 1.
    l = 0.5
    b = 0.1
    lc = 0.5
    J = 0.25
    g = 9.81

    Z[1,1] = 0                    # ∂θdot/∂θ
    Z[1,2] = 1                    # ∂θdot/∂ω
    Z[1,3] = 0                    # ∂θ/∂u
    Z[2,1] = -m*g*lc*cos(x[1])/J  # ∂ωdot/∂θ
    Z[2,2] = -b/J                 # ∂ωdot/∂ω
    Z[2,3] = 1/J                  # ∂ωdot/∂u
end

function pendulum_jacobian_params!(Z,x,u,p)
    Z[1,1] = 0                                    # ∂θdot/∂θ
    Z[1,2] = 1                                    # ∂θdot/∂ω
    Z[1,3] = 0                                    # ∂θ/∂u
    Z[2,1] = -p.m * p.g * p.lc * cos(x[1]) / p.J  # ∂ωdot/∂θ
    Z[2,2] = -p.b / p.J                           # ∂ωdot/∂ω
    Z[2,3] = 1/p.J                                # ∂ωdot/∂u
end

We can then pass these functions into the model instead of using ForwardDiff to calculate them

model = Model(pendulum_dynamics!, pendulum_jacobian!, n, m)
model = Model(pendulum_dynamics_params!, pendulum_jacobian_params!, n, m, params)

URDF Files

Instead of writing down the dynamics explicity, we can import the dynamics from geometry specified in a URDF model using RigidBodyDynamics.jl. Let's say we have a URDF file for a double pendulum and don't want to both writing down the dynamics, then we can create a model using any of the following methods

using RigidBodyDynamics
# From a string
urdf = "doublependulum.urdf"
model = Model(urdf)

# From a RigidBodyDynamics `Mechanism` type
mech = parse_urdf(urdf)  # return a Mechanism type
model = Model(mech)

Now let's say we want to control an acrobot, which can only control the first joint. We can pass in a vector of Booleans to specify which of the joints are "active."

joints = [true,false]

# From a string
urdf = "doublependulum.urdf"
model = Model(urdf,joints)

# From a RigidBodyDynamics `Mechanism` type
mech = parse_urdf(urdf)  # return a Mechanism type
model = Model(mech,joints)

A note on Model types

While the constructors look very similar, URDF models actually return a slightly different type than the the analytical ones of the first section. Analytical models are represented by

TrajectoryOptimization.AnalyticalModel — Type.

struct AnalyticalModel{D} <: Model{D}

Dynamics model

Holds all information required to uniquely describe a dynamic system, including a general nonlinear dynamics function of the form ẋ = f(x,u), where x ∈ ℜⁿ are the states and u ∈ ℜᵐ are the controls.

Dynamics function Model.f should be in the following forms: 'f!(ẋ,x,u)' and modify ẋ in place

source

whereas those created from a URDF are represented by

TrajectoryOptimization.RBDModel — Type.

struct RBDModel{D} <: Model{D}

RigidBodyDynamics model. Wrapper for a RigidBodyDynamics Mechanism

source

which explicitly stores the Mechanism internally.

Discrete Models

The previous methods all generate models with continuous dynamics (note that all of the models returned above will be of the type Model{Continuous}). In order to perform trajectory optimization we need to have discrete dynamics. Typically, we will form the continuous dynamics as we did above and then use a particular integration scheme to discretize it. Alternatively, we may know the analytical expression for the discrete dynamics.

From a continuous model

Assuming we have a model of type Model{Continuous}, we can discretize as follows:

model_discrete = Model{Discrete}(model,discretizer)

where discretizer is a function that returns a discretized version of the continuous dynamics. TrajectoryOptimization.jl offers the following integration schemes

midpoint

rk3 (Third Order Runge-Kutta)

rk4 (Fourth Order Runge-Kutta)

So to create a discrete model of the pendulum with fourth order Runge-Kutta integration we would do the following

# Create the continuous model (any of the previously mentioned method would work here)
params = (m=1, l=0.5, b=0.1, lc=0.5, J=0.25, g=9.81)
model = Model(pendulum_dynamics_params!, n, m, params)

# Discretize the continuous model
model_discrete = Model{Discrete}(model,rk4)

Analytical

An analytical model with discrete dynamics can be created using the following constructors

From Continuous Model

A discrete model can be created from a continuous model by specifying the integration (discretization) method. The following methods are currently supported

Use the following method to discretize a continuous model with one of the integration methods listed previously

TrajectoryOptimization.Model — Method.

Convert a continuous dynamics model into a discrete one using the given discretization function. The discretization function can either be one of the currently supported functions (midpoint, rk3, rk4) or a custom method that has the following form function discretizer(f::Function,dt::Float64) function fd!(xdot,x,u,dt) # Your code return nothing end return fd! end

source

API

The following constructors can be used to create a model from an analytic function, with or without parameters or analyical Jacobians

TrajectoryOptimization.Model — Type.

Model(f::Function, n::Int64, m::Int64) -> TrajectoryOptimization.AnalyticalModel{Continuous}
Model(f::Function, n::Int64, m::Int64, d::Dict{Symbol,Any}) -> TrajectoryOptimization.AnalyticalModel{Continuous}

Create a dynamics model, using ForwardDiff to generate the dynamics jacobian, with parameters Dynamics function passes in parameters: f(ẋ,x,u,p) where p in NamedTuple of parameters

source

TrajectoryOptimization.Model — Type.

Model(f::Function, n::Int64, m::Int64, p::NamedTuple) -> TrajectoryOptimization.AnalyticalModel{Continuous}
Model(f::Function, n::Int64, m::Int64, p::NamedTuple, d::Dict{Symbol,Any}) -> TrajectoryOptimization.AnalyticalModel{Continuous}

Create a dynamics model, using ForwardDiff to generate the dynamics jacobian, without parameters Dynamics function of the form: f(ẋ,x,u)

source

TrajectoryOptimization.Model — Type.

Model(f::Function, ∇f::Function, n::Int64, m::Int64, p::NamedTuple) -> TrajectoryOptimization.AnalyticalModel{Continuous}
Model(f::Function, ∇f::Function, n::Int64, m::Int64, p::NamedTuple, d::Dict{Symbol,Any}) -> TrajectoryOptimization.AnalyticalModel{Continuous}

Create a dynamics model with an analytical Jacobian, with parameters Dynamics functions pass in parameters: f(ẋ,x,u,p) ∇f(Z,x,u,p) where p in NamedTuple of parameters

source

TrajectoryOptimization.Model — Type.

Model(f::Function, ∇f::Function, n::Int64, m::Int64) -> TrajectoryOptimization.AnalyticalModel{Continuous}
Model(f::Function, ∇f::Function, n::Int64, m::Int64, d::Dict{Symbol,Any}) -> TrajectoryOptimization.AnalyticalModel{Continuous}

Create a dynamics model with an analytical Jacobian, without parameters Dynamics functions pass of the form: f(ẋ,x,u) ∇f(Z,x,u)

source

The following constructors can be used to create a Model from a URDF file

TrajectoryOptimization.Model — Method.

Model(mech, torques)

Model(mech::Mechanism, torques::Array{Bool, 1}) Constructor for an underactuated mechanism, where torques is a binary array that specifies whether a joint is actuated.

source

TrajectoryOptimization.Model — Method.

Model(mech)

Construct model from a Mechanism type from RigidBodyDynamics

source

TrajectoryOptimization.Model — Method.

Model(urdf)

Construct a fully actuated model from a string to a urdf file

source

TrajectoryOptimization.Model — Method.

Model(urdf, torques)

Construct a partially actuated model from a string to a urdf file, where torques is a binary array that specifies whether a joint is actuated.

source

evaluate!(ẋ::AbstractVector,model::Model,x,u)
evaluate!(Z::AbstractMatrix,ẋ::AbstractVector,model::Model,x,u)
jacobian!(Z::AbstractMatrix,ẋ::AbstractVector,model::Model,x,u)
jacobian!(Z::AbstractMatrix,model::Model,x,u)
evals(model::Model)
reset(model::Model)