Setting up a Dynamics Model

Setting up a Dynamics Model

Constraints

Constraint Type

Creating Constraints

Special Constraints

Setting up a Dynamics Model

Overview

Continuous Models

Analytical Models

URDF Models

Discrete Models

Analytical

From Continuous Model

API

Solvers

Iterative LQR (iLQR)

Setting up a Problem

Creating a Problem

Adding constraints

TrajectoryOptimization.jl

Overview

Getting Started

Creating a Model

Creating an Objective

Solving the Problem

Solve Methods

DIRCOL Method

Overview

The Model type holds information about the dynamics of the system. All dynamics are assumed to be state-space models of the system of the form ẋ = f(x,u) where ẋ is the state derivative, x an n-dimentional state vector, and u in an m-dimensional control input vector. The function f can be any nonlinear function.

TrajectoryOptimization.jl poses the trajectory optimization problem by discretizing the state and control trajectories, which requires discretizing the dynamics, turning the continuous time differential equation into a discrete time difference equation of the form x[k+1] = f(x[k],u[k]), where k is the time step. There many methods of performing this discretization, and TrajectoryOptimization.jl offers several of the most common methods.

Sometimes is it convenient to write down the difference equation directly, rather than running a differential equation through a discretizing integration method. TrajectoryOptimization.jl offers method deal directly with either continuous differential equations, or discrete difference equations.

The Model type is parameterized by the DynamicsType, which is either Continuous, or Discrete. The models holds the equation f and it's Jacobian, ∇f, along with the dimensions of the state and control vectors.

Models can be created by writing down the dynamics analytically or be generated from a URDF file via RigidBodyDynamics.jl.

Continuous Models

Continuous models assume differential equations are specified by an in-place function in one of the following forms:

f!(ẋ,x,u)
f!(ẋ,x,u,p)

and Jacobians of the form

∇f!(Z,x,u)
∇f!(Z,x,u,p)

where ẋ is the state derivative, p is a NamedTuple of model parameters, and Zis the (n × (n+m)) Jacobian matrix (i.e. [∇ₓf(x,u) ∇ᵤf(x,u)]). As soon as the model is created, however, only the forms without parameters (the top lines) are available. The Model type will automatically bake in the parameters. A new model must be created if a parameter is changed (this will be made easier in the future).

Analytical Models

TrajectoryOptimization.AnalyticalModel — Type.

struct AnalyticalModel{D} <: Model{D}

Dynamics model

Holds all information required to uniquely describe a dynamic system, including a general nonlinear dynamics function of the form ẋ = f(x,u), where x ∈ ℜⁿ are the states and u ∈ ℜᵐ are the controls.

Dynamics function Model.f should be in the following forms: 'f!(ẋ,x,u)' and modify ẋ in place

source

The following constructors can be used to create Continuous Analytical models

TrajectoryOptimization.Model — Type.

Model(f, n, m)
Model(f, n, m, d)

Create a dynamics model, using ForwardDiff to generate the dynamics jacobian, with parameters Dynamics function passes in parameters: f(ẋ,x,u,p) where p in NamedTuple of parameters

source

TrajectoryOptimization.Model — Type.

Model(f, n, m, p)
Model(f, n, m, p, d)

Create a dynamics model, using ForwardDiff to generate the dynamics jacobian, without parameters Dynamics function of the form: f(ẋ,x,u)

source

TrajectoryOptimization.Model — Type.

Model(f, ∇f, n, m, p)
Model(f, ∇f, n, m, p, d)

Create a dynamics model with an analytical Jacobian, with parameters Dynamics functions pass in parameters: f(ẋ,x,u,p) ∇f(Z,x,u,p) where p in NamedTuple of parameters

source

TrajectoryOptimization.Model — Type.

Model(f, ∇f, n, m)
Model(f, ∇f, n, m, d)

Create a dynamics model with an analytical Jacobian, without parameters Dynamics functions pass of the form: f(ẋ,x,u) ∇f(Z,x,u)

source

URDF Models

TrajectoryOptimization.RBDModel — Type.

struct RBDModel{D} <: Model{D}

RigidBodyDynamics model. Wrapper for a RigidBodyDynamics Mechanism

source

The following constructors can be used to create a Model from a URDF file

TrajectoryOptimization.Model — Method.

Model(mech, torques)

Model(mech::Mechanism, torques::Array{Bool, 1}) Constructor for an underactuated mechanism, where torques is a binary array that specifies whether a joint is actuated.

source

TrajectoryOptimization.Model — Method.

Model(mech)

Construct model from a Mechanism type from RigidBodyDynamics

source

TrajectoryOptimization.Model — Method.

Model(urdf)

Construct a fully actuated model from a string to a urdf file

source

TrajectoryOptimization.Model — Method.

Model(urdf, torques)

Construct a partially actuated model from a string to a urdf file, where torques is a binary array that specifies whether a joint is actuated.

source

Discrete Models

Discrete models assume difference equations are specified by an in-place function in one of the following forms:

f!(x′,x,u,dt)
f!(x′,x,u,p,dt)

and Jacobians of the form

∇f!(Z,x,u,dt)
∇f!(Z,x,u,p,dt)

where x′ is the state at the next time step, p is a NamedTuple of model parameters, and Zis the (n × (n+m)) Jacobian matrix (i.e. [∇ₓf(x,u) ∇ᵤf(x,u)]).

Analytical

An analytical model with discrete dynamics can be created using the following constructors

From Continuous Model

A discrete model can be created from a continuous model by specifying the integration (discretization) method. The following methods are currently supported

midpoint

rk3 (Third Order Runge-Kutta)

rk4 (Fourth Order Runge-Kutta)

Use the following method to discretize a continuous model with one of the integration methods listed previously

TrajectoryOptimization.Model — Method.

Convert a continuous dynamics model into a discrete one using the given discretization function. The discretization function can either be one of the currently supported functions (midpoint, rk3, rk4) or a custom method that has the following form function discretizer(f::Function,dt::Float64) function fd!(xdot,x,u,dt) # Your code return nothing end return fd! end

source

API

evaluate!(ẋ::AbstractVector,model::Model,x,u)
evaluate!(Z::AbstractMatrix,ẋ::AbstractVector,model::Model,x,u)
jacobian!(Z::AbstractMatrix,ẋ::AbstractVector,model::Model,x,u)
jacobian!(Z::AbstractMatrix,model::Model,x,u)
evals(model::Model)
reset(model::Model)