4. Setting up a Problem

4. Setting up a Problem

4. Setting up a Problem

Creating a Problem

Methods

Creating a Problem

TrajectoryOptimization.Problem — Type.

struct Problem{T<:AbstractFloat, D<:TrajectoryOptimization.DynamicsType}

Trajectory Optimization Problem. Contains the full definition of a trajectory optimization problem, including:

dynamics model (Model). Can be either continuous or discrete.

objective (Objective)

constraints (Constraints)

initial and final states

Primal variables (state and control trajectories)

Discretization information: knot points (N), time step (dt), and total time (tf)

Constructors:

Problem(model, obj X0, U0; integration, constraints, x0, xf, dt, tf, N)
Problem(model, obj, U0; integration, constraints, x0, xf, dt, tf, N)
Problem(model, obj; integration, constraints, x0, xf, dt, tf, N)

Arguments

model: Dynamics model. Can be either Discrete or Continuous

obj: Objective

X0: Initial state trajectory. If omitted it will be initialized with NaNs, to be later overwritten by the solver.

U0: Initial control trajectory. If omitted it will be initialized with zeros.

x0: Initial state. Defaults to zeros.

xf: Final state. Defaults to zeros.

dt: Time step

tf: Final time. Set to zero to specify a time penalized problem.

N: Number of knot points. Defaults to 51, unless specified by dt and tf.

integration: One of the defined integration types to discretize the continuous dynamics model. Defaults to :none, which will pass in the continuous dynamics (eg. for DIRCOL)

Both X0 and U0 can be either a Matrix or a Vector{Vector}, but must be the same. At least 2 of dt, tf, and N need to be specified (or just 1 of dt and tf).

source

Methods

TrajectoryOptimization.update_problem — Function.

update_problem(prob; kwargs...)

Create a new problem from another, specifing all fields as keyword arguments The newProb argument can be set to true if a the primal variables are to be copied, otherwise they will be passed to the modified problem.

source

TrajectoryOptimization.initial_controls! — Function.

initial_controls!(prob, U0)

Set the initial control trajectory for a problem. U0 can be either a Matrix or a Vector{Vector}

source

TrajectoryOptimization.initial_states! — Function.

Set the initial state trajectory for a problem. X0 can be either a Matrix or a Vector{Vector}

source

TrajectoryOptimization.set_x0! — Function.

set_x0!(prob, x0)

Set the initial state

source

Base.size — Method.

n,m,N = size(p::Problem)

Return the number of states (n), number of controls (m), and the number of knot points (N)

source

Base.copy — Method.

copy(p::Problem) -> Problem{_1,_2} where _2 where _1

Copy a problem

source

TrajectoryOptimization.is_constrained — Function.

Checks if a problem has any constraints

source

TrajectoryOptimization.max_violation — Method.

Evaluate the maximum contraint violation

source

Evaluate maximum constraint violation

source

TrajectoryOptimization.final_time — Function.

final_time(prob)

Get the total time for tje trajectory (applicable for time-penalized problems)

source