Problem

Problem

Definition

Methods

Definition

TrajectoryOptimization.Problem — Type

struct Problem{T<:AbstractFloat}

Trajectory Optimization Problem. Contains the full definition of a trajectory optimization problem, including:

dynamics model (Model)

objective (Objective)

constraints (ConstraintSet)

initial and final states

Primal variables (state and control trajectories)

Discretization information: knot points (N), time step (dt), and total time (tf)

Constructors:

Problem(model, obj, constraints, x0, xf, Z, N, tf)
Problem(model, obj, x0, tf; xf, constraints, N, X0, U0, dt, integration)

where Z is a trajectory (Vector of KnotPoints)

Arguments

model: A DiscreteDynamics model. If a ContinuousDynamics model is provided, it will be converted to a DiscretizedDynamics model via the integrator specified by the integration keyword argument.

obj: Objective

X0: Initial state trajectory. If omitted it will be initialized with NaNs, to be later overwritten by the solver.

U0: Initial control trajectory. If omitted it will be initialized with zeros.

x0: Initial state. Defaults to zeros.

xf: Final state. Defaults to zeros.

dt: Time step

tf: Final time. Set to zero to specify a time penalized problem.

N: Number of knot points. Defaults to 51, unless specified by dt and tf.

integration: One of the defined integration types to discretize the continuous dynamics model.

Both X0 and U0 can be either a Matrix or a Vector{Vector}, but must be the same. At least 2 of dt, tf, and N need to be specified (or just 1 of dt and tf).

Methods

Getters

TrajectoryOptimization.cost — Method

cost(::Problem)

Compute the cost for the current trajectory

RobotDynamics.states — Method

states(::Problem)

Get the state trajectory.

RobotDynamics.controls — Method

controls(::Problem)

Get the control trajectory

Base.size — Method

size(A::AbstractArray, [dim])

Return a tuple containing the dimensions of A. Optionally you can specify a dimension to just get the length of that dimension.

Note that size may not be defined for arrays with non-standard indices, in which case axes may be useful. See the manual chapter on arrays with custom indices.

Examples

julia> A = fill(1, (2,3,4));

julia> size(A)
(2, 3, 4)

julia> size(A, 2)
3

TrajectoryOptimization.num_constraints — Method

Get the number of constraint values at each time step

TrajectoryOptimization.get_constraints — Function

Get problem constraints. Returns AbstractConstraintSet.

TrajectoryOptimization.get_model — Function

Get the dynamics model. Returns RobotDynamics.AbstractModel.

TrajectoryOptimization.get_trajectory — Function

Get the trajectory. Returns an RobotDynamics.AbstractTrajectory

TrajectoryOptimization.get_initial_state — Function

Get the in initial state. Returns an AbstractVector.

Missing docstring.

Missing docstring for RobotDynamics.get_times(::Problem). Check Documenter's build log for details.

Missing docstring.

Missing docstring for integration(::Problem). Check Documenter's build log for details.

TrajectoryOptimization.is_constrained — Function

Determines if the problem is constrained.

Setters

TrajectoryOptimization.initial_controls! — Method

initial_controls!(::Problem, U0::Vector{<:AbstractVector})
initial_controls!(::Problem, U0::AbstractMatrx)

Copy the control trajectory

TrajectoryOptimization.initial_states! — Method

initial_states!(::Problem, X0::Vector{<:AbstractVector})
initial_states!(::Problem, X0::AbstractMatrix)

Copy the state trajectory

TrajectoryOptimization.initial_trajectory! — Function

initial_trajectory!(prob::Problem, Z)

Copy the trajectory

TrajectoryOptimization.set_initial_state! — Method

set_initial_state!(prob::Problem, x0::AbstractVector)

Set the initial state in prob to x0

TrajectoryOptimization.set_initial_time! — Method

set_initial_time!(prob, t0)

Set the initial time of the optimization problem, shifting the time of all points in the trajectory. Returns the updated final time.

TrajectoryOptimization.set_goal_state! — Function

set_goal_state!(prob::Problem, xf::AbstractVector; objective=true, constraint=true)

Change the goal state. If the appropriate flags are true, it will also modify a GoalConstraint and the objective, assuming it's an LQRObjective.

TrajectoryOptimization.change_integration — Function

change_integration(prob::Problem, Q<:QuadratureRule)

Change dynamics integration for the problem. Returns a new problem.

Other

RobotDynamics.rollout! — Function

rollout!(::Problem)

Simulate the dynamics forward from the initial condition x0 using the controls in the trajectory Z. If a problem is passed in, Z = prob.Z, model = prob.model, and x0 = prob.x0.

Base.copy — Method

Copy the problem