Solvers

Solvers

Solvers

Iterative LQR (iLQR)

Augmented Lagrangian

ALTRO

Direct Collocation (DIRCOL)

Iterative LQR (iLQR)

iLQR is an unconstrained indirect method for trajectory optimization that parameterizes only the controls and enforces strict dynamics feasibility at every iteration by simulating forward the dynamics with an LQR feedback controller. The main algorithm consists of two parts: 1) a backward pass that uses Differential Dynamic Programming to compute recursively a quadratic approximation of the cost-to-go, along with linear feedback and feed-forward gain matrices, K and d, respectively, for an LQR tracking controller, and 2) a forward pass that uses the gains K and d to simulate forward the full nonlinear dynamics with feedback.

The iLQR solver has the following solver options

TrajectoryOptimization.iLQRSolverOptions — Type.

mutable struct iLQRSolverOptions{T} <: TrajectoryOptimization.AbstractSolverOptions{T}

Solver options for the iterative LQR (iLQR) solver. iLQR is an indirect, unconstrained solver.

verbose

Print summary at each iteration Default: false

live_plotting

Live plotting Default: :off

cost_tolerance

dJ < ϵ, cost convergence criteria for unconstrained solve or to enter outerloop for constrained solve Default: 0.0001

gradient_type

gradient type: :todorov, :feedforward Default: :todorov

gradient_norm_tolerance

gradient_norm < ϵ, gradient norm convergence criteria Default: 1.0e-5

iterations

iLQR iterations Default: 300

dJ_counter_limit

restricts the total number of times a forward pass fails, resulting in regularization, before exiting Default: 10

square_root

use square root method backward pass for numerical conditioning Default: false

line_search_lower_bound

forward pass approximate line search lower bound, 0 < linesearchlowerbound < linesearchupperbound Default: 1.0e-8

line_search_upper_bound

forward pass approximate line search upper bound, 0 < linesearchlowerbound < linesearchupperbound < ∞ Default: 10.0

iterations_linesearch

maximum number of backtracking steps during forward pass line search Default: 20

bp_reg_initial

initial regularization Default: 0.0

bp_reg_increase_factor

regularization scaling factor Default: 1.6

bp_reg_max

maximum regularization value Default: 1.0e8

bp_reg_min

minimum regularization value Default: 1.0e-8

bp_reg_type

type of regularization- control: () + ρI, state: (S + ρI); see Synthesis and Stabilization of Complex Behaviors through Online Trajectory Optimization Default: :control

bp_reg_fp

additive regularization when forward pass reaches max iterations Default: 10.0

bp_sqrt_inv_type

type of matrix inversion for bp sqrt step Default: :pseudo

bp_reg_sqrt_initial

initial regularization for square root method Default: 1.0e-6

bp_reg_sqrt_increase_factor

regularization scaling factor for square root method Default: 10.0

max_cost_value

maximum cost value, if exceded solve will error Default: 1.0e8

max_state_value

maximum state value, evaluated during rollout, if exceded solve will error Default: 1.0e8

max_control_value

maximum control value, evaluated during rollout, if exceded solve will error Default: 1.0e8

source

Augmented Lagrangian

Augmented Lagrangian

TrajectoryOptimization.AugmentedLagrangianSolverOptions — Type.

mutable struct AugmentedLagrangianSolverOptions{T} <: TrajectoryOptimization.AbstractSolverOptions{T}

Solver options for the augmented Lagrangian solver. Augmented Lagrangian is a general method for solving constrained problems by solving a sequence of unconstrained problems.

verbose

Print summary at each iteration Default: false

opts_uncon

unconstrained solver options Default: iLQRSolverOptions{T}()

cost_tolerance

dJ < ϵ, cost convergence criteria for unconstrained solve or to enter outerloop for constrained solve Default: 0.0001

cost_tolerance_intermediate

dJ < ϵ_int, intermediate cost convergence criteria to enter outerloop of constrained solve Default: 0.001

gradient_norm_tolerance

gradient_norm < ϵ, gradient norm convergence criteria Default: 1.0e-5

gradient_norm_tolerance_intermediate

gradientnormint < ϵ, gradient norm intermediate convergence criteria Default: 1.0e-5

constraint_tolerance

max(constraint) < ϵ, constraint convergence criteria Default: 0.001

constraint_tolerance_intermediate

max(constraint) < ϵ_int, intermediate constraint convergence criteria Default: 0.001

iterations

maximum outerloop updates Default: 30

dual_min

minimum Lagrange multiplier Default: -1.0e8

dual_max

maximum Lagrange multiplier Default: 1.0e8

penalty_max

maximum penalty term Default: 1.0e8

penalty_initial

initial penalty term Default: 1.0

penalty_scaling

penalty update multiplier; penalty_scaling > 0 Default: 10.0

penalty_scaling_no

penalty update multiplier when μ should not be update, typically 1.0 (or 1.0 + ϵ) Default: 1.0

constraint_decrease_ratio

ratio of current constraint to previous constraint violation; 0 < constraintdecreaseratio < 1 Default: 0.25

outer_loop_update_type

type of outer loop update (default, feedback) Default: :default

active_constraint_tolerance

numerical tolerance for constraint violation Default: 0.0

source

ALTRO

ALTRO

ALTROSolverOptions

Direct Collocation (DIRCOL)